#189 extract GraphQLHandler to common
