+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/stores/webservice/Setup.kt
@@ -3,6 +3,6 @@ package app.ehrenamtskarte.backend.stores.webservice
 import io.javalin.Javalin
 
 fun setupApp(app: Javalin) {
-    val graphQLHandler = GraphQLHandler()
+    val graphQLHandler = StoresGraphQLHandler()
     app.post("/") { ctx -> graphQLHandler.handle(ctx) }
 }
