+++ backend/src/main/kotlin/app/ehrenamtskarte/backend/stores/webservice/StoresGraphQLHandler.kt
@@ -1,0 +1,24 @@
+package app.ehrenamtskarte.backend.stores.webservice
+
+import app.ehrenamtskarte.backend.common.webservice.GraphQLHandler
+import app.ehrenamtskarte.backend.stores.webservice.schema.AcceptingStoreQueryService
+import app.ehrenamtskarte.backend.stores.webservice.schema.CategoriesQueryService
+import com.expediagroup.graphql.SchemaGeneratorConfig
+import com.expediagroup.graphql.TopLevelObject
+import com.expediagroup.graphql.toSchema
+import graphql.GraphQL
+
+class StoresGraphQLHandler : GraphQLHandler(
+    ::initializeDataLoaderRegistry,
+    graphQL
+) {
+    companion object {
+        private val config = SchemaGeneratorConfig(supportedPackages = listOf("app.ehrenamtskarte.backend.stores.webservice.schema"))
+        private val queries = listOf(
+            TopLevelObject(AcceptingStoreQueryService()),
+            TopLevelObject(CategoriesQueryService())
+        )
+        val graphQLSchema = toSchema(config, queries)
+        val graphQL = GraphQL.newGraphQL(graphQLSchema).build()!!
+    }
+}
